Finished reorganization of directory
